84 user management and access control UI

+++ src/controller/backend/Endpoints/UserManagementEndpoints.cs
@@ -1,6 +1,9 @@
-﻿using ApiSchema.Usermanagement;
+﻿using System.Security.Claims;
+using ApiSchema.Enums;
+using ApiSchema.Usermanagement;
 using Backend.Services.Identity;
 using Data.System;
+using LanguageExt.ClassInstances;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -23,15 +26,50 @@ public static class UserManagementEndpoints
                 .Include(s => s.UserProfile).FirstOrDefaultAsync(s => s.UserName == user);
             if (userEntity is null)
                 return Results.NotFound();
-            return Results.Ok(new UsersModel()
+            var userPermissions = await userManager.GetClaimsAsync(userEntity);
+            return Results.Ok(new UsersModel
             {
                 Username = userEntity.UserName!,
                 CanChangeUsername = userEntity.CanChangeUsername,
                 IsActive = userEntity.IsActive,
-                Language = (int)userEntity.UserProfile!.Language,
-                UseDarkMode = userEntity.UserProfile.UseDarkMode
+                Language = userEntity.UserProfile!.Language.ToString(),
+                UseDarkMode = userEntity.UserProfile.UseDarkMode,
+                Permissions = userPermissions.Select(s => s.Value).ToList(),
             });
         }).RequirePermission(UserClaims.UserViewUsers);
+
+        group.MapPost("/users/{user}", async (
+            UserManager<SystemUser> userManager,
+            [FromServices] SystemDbContext dbContext,
+            string user,
+            [FromBody] UsersModel model) =>
+        {
+            var errorMessage = "";
+            var userEntity = await dbContext.Users
+                .Include(s => s.UserProfile).FirstOrDefaultAsync(s => s.UserName == user);
+            if(user != model.Username && await dbContext.Users.AnyAsync(s => s.UserName == model.Username))
+                errorMessage += "Username already exists";
+            if (userEntity is null)
+                return Results.NotFound();
+            if(userEntity.CanChangeUsername)
+                userEntity.UserName = model.Username;
+            if (userEntity.UserName != "admin")
+                userEntity.CanChangeUsername = model.CanChangeUsername;
+            userEntity.IsActive = model.IsActive;
+            var claims = await userManager.GetClaimsAsync(userEntity);
+            claims.ToList().ForEach(c => userManager.RemoveClaimAsync(userEntity, c));
+            model.Permissions.ForEach(p => userManager.AddClaimAsync(userEntity, new Claim("Permissions", p)));
+            var parsable = Enum.TryParse<Language>(model.Language.ToString(), out var language);
+            if(parsable)
+                userEntity.UserProfile!.Language = language;
+            else
+                errorMessage += errorMessage == "" ? "Language not found" : ", Language not found";
+            userEntity.UserProfile!.UseDarkMode = model.UseDarkMode;
+            if(errorMessage != "")
+                return Results.BadRequest(errorMessage);
+            await dbContext.SaveChangesAsync();
+            return Results.Ok();
+        }).RequirePermission(UserClaims.UserChangeUserUsername);
 
         group.MapPost("/users/{user}/username", async (
             [FromRoute] string user,
@@ -43,7 +81,7 @@ public static class UserManagementEndpoints
             if (userEntity is null)
                 return Results.NotFound("User not found");
             if (!userEntity.CanChangeUsername)
-                return Results.Forbid();
+                return Results.Unauthorized();
             userEntity.UserName = model.Username;
             var result = await context.SaveChangesAsync();
             return result == 1 ? Results.Ok() : Results.InternalServerError();
@@ -58,7 +96,7 @@ public static class UserManagementEndpoints
             if (userEntity is null)
                 return Results.NotFound("User not found");
             await userManager.RemovePasswordAsync(userEntity);
-            var result = await userManager.AddPasswordAsync(userEntity, model.NewPassword);
+            var result = await userManager.AddPasswordAsync(userEntity, model.NewPassword!);
             return result == IdentityResult.Success ? Results.Ok() : Results.BadRequest(result.Errors);
         }).RequirePermission(UserClaims.UserChangeUserPassword);
 
@@ -107,21 +145,26 @@ public static class UserManagementEndpoints
 
         group.MapGet("/users", async (
                 HttpContext context,
-                UserManager<SystemUser> userManager) =>
+                UserManager<SystemUser> userManager,
+                RoleManager<IdentityRole> roleManager) =>
             {
                 var user = await userManager.GetUserAsync(context.User);
                 if (user is null)
                     return Results.Unauthorized();
                 var users = await userManager.Users.Include(s => s.UserProfile).ToListAsync();
-                var usersInAdmin = await userManager.GetUsersInRoleAsync("admin");
-                return Results.Ok(users.Select(u => new UsersModel()
+                Dictionary<string, List<string>> permissions = new Dictionary<string, List<string>>();
+                users.ForEach(u => permissions.Add(u.UserName!, userManager.GetClaimsAsync(u).Result.Select(s => s.Value).ToList()));
+                List<UsersModel> allUsers = new();
+                users.ForEach(u => allUsers.Add(new()
                 {
                     Username = u.UserName!,
-                    Language = (int) u.UserProfile!.Language,
-                    UseDarkMode = u.UserProfile.UseDarkMode,
                     CanChangeUsername = u.CanChangeUsername,
                     IsActive = u.IsActive,
+                    Language = u.UserProfile!.Language.ToString(),
+                    UseDarkMode = u.UserProfile.UseDarkMode,
+                    Permissions = permissions[u.UserName!]
                 }));
+                return Results.Ok(allUsers);
             }).RequirePermission(UserClaims.UserViewUsers);
 
         group.MapPost("/users", async (
@@ -133,9 +176,6 @@ public static class UserManagementEndpoints
             var user = await userManager.GetUserAsync(context.User);
             if (user is null)
                 return Results.Unauthorized();
-            var usersInAdmin = await userManager.GetUsersInRoleAsync("admin");
-            if (!usersInAdmin.Contains(user))
-                return Results.Forbid();
             var newUser = await userManager.FindByNameAsync(model.Username);
             if (newUser is not null)
                 return Results.BadRequest("User already exists");
@@ -152,7 +192,7 @@ public static class UserManagementEndpoints
                 return Results.InternalServerError();
 
 
-            return Results.Created();
+            return Results.Created(user.UserName, user);
         }).RequirePermission(UserClaims.UserAddUser);
 
         group.MapDelete("/users/{user}", async (
@@ -166,10 +206,12 @@ public static class UserManagementEndpoints
             var usersInAdmin = await userManager.GetUsersInRoleAsync("admin");
             if (!usersInAdmin.Contains(userEntity))
                 return Results.Forbid();
-            var username = context.Request.Query["username"];
             var userToRemove = await userManager.FindByNameAsync(user);
             if (userToRemove is null)
                 return Results.NotFound();
+            List<string> blockesUsernames = new () { "admin", "Visitor" };
+            if (blockesUsernames.Contains(userToRemove.UserName!))
+                return Results.Unauthorized();
             var result = await userManager.DeleteAsync(userToRemove);
             return result.Succeeded ? Results.Ok() : Results.InternalServerError();
         }).RequirePermission(UserClaims.UserRemoveUser);
